Organize code into packages. Add ShoppingItem class.
