+++ app/src/main/java/com/lambdaschool/sprint2_challenge/model/ShoppingItem.kt
@@ -1,0 +1,3 @@
+package com.lambdaschool.sprint2_challenge.model
+
+class ShoppingItem (val itemName: String)
